Interface simplification (#40)
