+++ src/main/java/io/github/jpmorganchase/fusion/api/context/MultipartTransferContext.java
@@ -1,5 +1,6 @@
 package io.github.jpmorganchase.fusion.api.context;
 
+import static io.github.jpmorganchase.fusion.api.context.MultipartTransferContext.MultipartTransferStatus.INITIATED;
 import static io.github.jpmorganchase.fusion.api.context.MultipartTransferContext.MultipartTransferStatus.TRANSFERRED;
 
 import io.github.jpmorganchase.fusion.api.response.UploadedPart;
@@ -19,14 +20,14 @@ public class MultipartTransferContext {
     MultipartTransferStatus status;
     final List<UploadedPartContext> parts;
     int chunkSize;
-    int totalBytes;
+    long totalBytes;
     int totalPartsCount;
 
     public static MultipartTransferContext started(Operation operation) {
         return new MultipartTransferContext(operation, MultipartTransferStatus.INITIATED, new ArrayList<>(), 0, 0, 0);
     }
 
-    public MultipartTransferContext transferred(int chunkSize, int totalBytes, int totalPartsCount) {
+    public MultipartTransferContext transferred(int chunkSize, long totalBytes, int totalPartsCount) {
         this.parts.sort(Comparator.comparingInt(UploadedPartContext::getPartNo));
         this.status = TRANSFERRED;
         this.chunkSize = chunkSize;
@@ -65,12 +66,11 @@ public class MultipartTransferContext {
     }
 
     public boolean canProceedToComplete() {
-        // TODO : knighto : This is one for the future once better re-try logic has been implemented
-        return TRANSFERRED.equals(this.status) && parts.size() > 0;
+        return (this.status == TRANSFERRED) && (parts.size() > 0);
     }
 
     public boolean canProceedToTransfer() {
-        return MultipartTransferStatus.INITIATED.equals(status);
+        return this.status == INITIATED;
     }
 
     public UploadedParts uploadedParts() {
