+++ src/main/java/io/github/jpmorganchase/fusion/http/JdkClient.java
@@ -2,16 +2,16 @@ package io.github.jpmorganchase.fusion.http;
 
 import java.io.*;
 import java.lang.invoke.MethodHandles;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.Proxy;
-import java.net.URL;
+import java.net.*;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
+import java.util.Objects;
 import java.util.function.Function;
+import lombok.Builder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+@Builder
 public class JdkClient implements Client {
 
     private static final Logger logger =
@@ -21,14 +21,6 @@ public class JdkClient implements Client {
     public static final String METHOD_PUT = "PUT";
     public static final String METHOD_DELETE = "DELETE";
     private final Proxy proxy;
-
-    public JdkClient(Proxy proxy) {
-        this.proxy = proxy;
-    }
-
-    public JdkClient() {
-        this.proxy = Proxy.NO_PROXY;
-    }
 
     @Override
     public HttpResponse<String> get(String path, Map<String, String> headers) {
@@ -181,5 +173,43 @@ public class JdkClient implements Client {
                 .addArgument(method)
                 .addArgument(connection.getURL())
                 .log();
+    }
+
+    public static JdkClientBuilder builder() {
+        return new CustomJdkClientBuilder();
+    }
+
+    public static class JdkClientBuilder {
+
+        String url;
+        int port;
+        Proxy proxy = Proxy.NO_PROXY;
+
+        public JdkClientBuilder url(String url) {
+            this.url = url;
+            return this;
+        }
+
+        public JdkClientBuilder port(int port) {
+            this.port = port;
+            return this;
+        }
+
+        public JdkClientBuilder noProxy() {
+            this.proxy = Proxy.NO_PROXY;
+            return this;
+        }
+    }
+
+    private static class CustomJdkClientBuilder extends JdkClientBuilder {
+        @Override
+        public JdkClient build() {
+
+            if (Objects.nonNull(url)) {
+                this.proxy = new Proxy(Proxy.Type.HTTP, new InetSocketAddress(url, port));
+            }
+
+            return super.build();
+        }
     }
 }
