+++ src/main/java/io/github/jpmorganchase/fusion/oauth/provider/FusionTokenProvider.java
@@ -1,0 +1,3 @@
+package io.github.jpmorganchase.fusion.oauth.provider;
+
+public interface FusionTokenProvider extends SessionTokenProvider, DatasetTokenProvider {}
