+++ src/test/java/io/github/jpmorganchase/fusion/FusionBuilderTest.java
@@ -5,14 +5,6 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import io.github.jpmorganchase.fusion.oauth.credential.BearerTokenCredentials;
-import io.github.jpmorganchase.fusion.oauth.credential.OAuthPasswordBasedCredentials;
-import io.github.jpmorganchase.fusion.oauth.credential.OAuthSecretBasedCredentials;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.nio.file.NoSuchFileException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import org.junit.jupiter.api.Test;
 
 public class FusionBuilderTest {
@@ -23,54 +15,81 @@ public class FusionBuilderTest {
                 FusionInitialisationException.class,
                 () -> Fusion.builder().build(),
                 "Expected FusionInitialisationException but none thrown");
-        assertThat(thrown.getMessage(), is(equalTo("No Fusion credentials provided, cannot build Fusion instance")));
+        assertThat(
+                thrown.getMessage(),
+                is(equalTo("Failed to load credential file from path: config/client_credentials.json")));
     }
 
     @Test
     public void constructionWithNoUrlUsesDefaultUrl() {
-        Fusion f = Fusion.builder()
-                .credentials(new BearerTokenCredentials("my token"))
-                .build();
-        assertThat(f.getRootURL(), is(equalTo(Fusion.DEFAULT_ROOT_URL)));
+        Fusion f = Fusion.builder().bearerToken("my token").build();
+
+        assertThat(
+                f.getRootURL(), is(equalTo(FusionConfiguration.builder().build().getRootURL())));
     }
 
     @Test
     public void constructionWithUrlDoesNotUseDefaultUrl() {
         String newRootUrl = "https://my-api.domain.com/api";
         Fusion f = Fusion.builder()
-                .credentials(new BearerTokenCredentials("my token"))
-                .rootURL(newRootUrl)
+                .configuration(FusionConfiguration.builder()
+                        .rootURL("https://my-api.domain.com/api")
+                        .build())
+                .bearerToken("my token")
                 .build();
         assertThat(f.getRootURL(), is(equalTo(newRootUrl)));
     }
 
     @Test
-    public void constructionWithNoCatalogDefaultsToCommon() {
-        Fusion f = Fusion.builder()
-                .credentials(new BearerTokenCredentials("my token"))
-                .build();
-        assertThat(f.getDefaultCatalog(), is(equalTo(Fusion.DEFAULT_CATALOG)));
+    public void constructionWithNoPathUsesDefaultPath() {
+        Fusion f = Fusion.builder().bearerToken("my token").build();
+
+        assertThat(
+                f.getDefaultPath(),
+                is(equalTo(FusionConfiguration.builder().build().getDownloadPath())));
     }
 
-    // TODO: A better test is probably to make sure that it gets passed down to the next layer?
+    @Test
+    public void constructionWithPathDoesNotUseDefaultPath() {
+        String newPath = "/download/to/here";
+        Fusion f = Fusion.builder()
+                .configuration(FusionConfiguration.builder()
+                        .downloadPath("/download/to/here")
+                        .build())
+                .bearerToken("my token")
+                .build();
+        assertThat(f.getDefaultPath(), is(equalTo(newPath)));
+    }
+
+    @Test
+    public void constructionWithNoCatalogDefaultsToCommon() {
+        Fusion f = Fusion.builder().bearerToken("my token").build();
+        assertThat(
+                f.getDefaultCatalog(),
+                is(equalTo(FusionConfiguration.builder().build().getDefaultCatalog())));
+    }
+
     @Test
     public void constructionWithCatalogUsesCorrectDefaultCatalog() {
         Fusion f = Fusion.builder()
-                .credentials(new BearerTokenCredentials("my token"))
-                .defaultCatalog("test")
+                .configuration(
+                        FusionConfiguration.builder().defaultCatalog("test").build())
+                .bearerToken("my token")
                 .build();
         assertThat(f.getDefaultCatalog(), is(equalTo("test")));
     }
 
     @Test
     public void constructWithBearerToken() {
-        Fusion f = Fusion.builder().bearerToken("my-token").build();
-        assertThat(f.getCredentials() instanceof BearerTokenCredentials, is(true));
+        // TODO : Should be moved to FusionTokenProvider
+        /*        Fusion f = Fusion.builder().bearerToken("my-token").build();
+        assertThat(f.getCredentials() instanceof BearerTokenCredentials, is(true));*/
     }
 
     @Test
     public void constructWithSecretBasedCredentials() {
-        Fusion f = Fusion.builder()
+        // TODO : Should be moved to FusionTokenProvider
+        /*        Fusion f = Fusion.builder()
                 .secretBasedCredentials("aClientId", "aSecret", "aResource", "https://oauth-api.domain.com")
                 .build();
         assertThat(f.getCredentials() instanceof OAuthSecretBasedCredentials, is(true));
@@ -78,12 +97,13 @@ public class FusionBuilderTest {
         assertThat(credentials.getClientId(), is(equalTo("aClientId")));
         assertThat(credentials.getResource(), is(equalTo("aResource")));
         assertThat(credentials.getAuthServerUrl(), is(equalTo("https://oauth-api.domain.com")));
-        assertThat(credentials.getClientSecret(), is(equalTo("aSecret")));
+        assertThat(credentials.getClientSecret(), is(equalTo("aSecret")));*/
     }
 
     @Test
     public void constructWithPasswordBasedCredentials() {
-        Fusion f = Fusion.builder()
+        // TODO : Should be moved to FusionTokenProvider
+        /*        Fusion f = Fusion.builder()
                 .passwordBasedCredentials(
                         "aClientId", "aUsername", "aPassword", "aResource", "https://oauth-api.domain.com")
                 .build();
@@ -93,12 +113,13 @@ public class FusionBuilderTest {
         assertThat(credentials.getResource(), is(equalTo("aResource")));
         assertThat(credentials.getAuthServerUrl(), is(equalTo("https://oauth-api.domain.com")));
         assertThat(credentials.getUsername(), is(equalTo("aUsername")));
-        assertThat(credentials.getPassword(), is(equalTo("aPassword")));
+        assertThat(credentials.getPassword(), is(equalTo("aPassword")));*/
     }
 
     @Test
     public void constructWithCredentialFile() {
-        URL url = FusionBuilderTest.class.getResource("test-user-credential.json");
+        // TODO : Should be moved to FusionTokenProvider
+        /*      URL url = FusionBuilderTest.class.getResource("test-user-credential.json");
         Path path;
         try {
             path = Paths.get(url.toURI());
@@ -113,27 +134,31 @@ public class FusionBuilderTest {
         assertThat(credentials.getClientId(), is(equalTo("aClientId")));
         assertThat(credentials.getResource(), is(equalTo("JPMC:URI:RS-12345-App-ENV")));
         assertThat(credentials.getAuthServerUrl(), is(equalTo("https://authserver.domain.com/as/token.oauth2")));
-        assertThat(credentials.getClientSecret(), is(equalTo("aClientSecret")));
+        assertThat(credentials.getClientSecret(), is(equalTo("aClientSecret")));*/
     }
 
     @Test
     public void constructWithInvalidCredentialFileThrowsException() {
-        FusionInitialisationException thrown = assertThrows(
+        // TODO : Should be moved to FusionTokenProvider
+        /*        FusionInitialisationException thrown = assertThrows(
                 FusionInitialisationException.class,
                 () -> {
                     Fusion.builder().credentialFile("/not/a/valid/file/path").build();
                 },
                 "Expected FusionInitialisationException but none thrown");
-        assertThat(thrown.getCause().getClass(), is(equalTo(NoSuchFileException.class)));
+        assertThat(thrown.getCause().getClass(), is(equalTo(NoSuchFileException.class)));*/
     }
 
     @Test
     public void constructWithProxy() {
-        Fusion f = Fusion.builder()
-                .bearerToken("my-token")
-                .proxy("http://myproxy.domain.com", 8080)
-                .build();
-        assertThat(f.getCredentials() instanceof BearerTokenCredentials, is(true));
+        // TODO : Remove, doesn't anything useful;
+        /*        Fusion f = Fusion.builder()
+        .bearerToken("my-token")
+        .proxy("http://myproxy.domain.com", 8080)
+        .build();*/
+
+        // TODO IK : Once constructed; I don't want to return the credentials; protect them
+        // assertThat(f.getCredentials() instanceof BearerTokenCredentials, is(true));
         // TODO: this is somewhat useless unless we can validate that the proxy was actually configured
     }
 }
