+++ src/test/java/io/github/jpmorganchase/fusion/FusionTest.java
@@ -6,9 +6,7 @@ import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
-import io.github.jpmorganchase.fusion.api.APIDownloader;
 import io.github.jpmorganchase.fusion.api.APIManager;
-import io.github.jpmorganchase.fusion.api.APIUploader;
 import io.github.jpmorganchase.fusion.http.Client;
 import io.github.jpmorganchase.fusion.model.*;
 import io.github.jpmorganchase.fusion.oauth.credential.BearerTokenCredentials;
@@ -37,16 +35,12 @@ public class FusionTest {
     private APIManager apiManager;
 
     @Mock
-    private APIUploader apiUploader;
-
-    @Mock
-    private APIDownloader apiDownloader;
-
-    @Mock
     private APIResponseParser responseParser;
 
     @Mock
     private Client httpClient;
+
+    private final FusionConfiguration config = FusionConfiguration.builder().build();
 
     @Test
     public void testListDatasetsInteraction() throws Exception {
@@ -73,7 +67,7 @@ public class FusionTest {
         Map<String, Dataset> stubResponse = new HashMap<>();
         stubResponse.put("first", Dataset.builder().identifier("dataset1").build());
 
-        when(apiManager.callAPI(String.format("%1scatalogs/%2s/datasets", Fusion.DEFAULT_ROOT_URL, catalog)))
+        when(apiManager.callAPI(String.format("%1scatalogs/%2s/datasets", config.getRootURL(), catalog)))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseDatasetResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -105,7 +99,7 @@ public class FusionTest {
         Map<String, DataProduct> stubResponse = new HashMap<>();
         stubResponse.put("first", DataProduct.builder().identifier("product1").build());
 
-        when(apiManager.callAPI(String.format("%1scatalogs/%2s/products", Fusion.DEFAULT_ROOT_URL, catalog)))
+        when(apiManager.callAPI(String.format("%1scatalogs/%2s/products", config.getRootURL(), catalog)))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseDataProductResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -120,8 +114,7 @@ public class FusionTest {
         stubResponse.put("first", DatasetSeries.builder().identifier("dataset1").build());
 
         when(apiManager.callAPI(String.format(
-                        "%1scatalogs/%2s/datasets/%3s/datasetseries",
-                        Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset")))
+                        "%1scatalogs/%2s/datasets/%3s/datasetseries", config.getRootURL(), "common", "sample_dataset")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseDatasetSeriesResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -137,8 +130,7 @@ public class FusionTest {
         stubResponse.put("first", Attribute.builder().identifier("attribute1").build());
 
         when(apiManager.callAPI(String.format(
-                        "%1scatalogs/%2s/datasets/%3s/attributes",
-                        Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset")))
+                        "%1scatalogs/%2s/datasets/%3s/attributes", config.getRootURL(), "common", "sample_dataset")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseAttributeResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -157,8 +149,7 @@ public class FusionTest {
         stubResponse.put("attribute1", attribute1);
 
         when(apiManager.callAPI(String.format(
-                        "%1scatalogs/%2s/datasets/%3s/attributes",
-                        Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset")))
+                        "%1scatalogs/%2s/datasets/%3s/attributes", config.getRootURL(), "common", "sample_dataset")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseResourcesUntyped("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -176,7 +167,7 @@ public class FusionTest {
 
         when(apiManager.callAPI(String.format(
                         "%1scatalogs/%2s/datasets/%3s/datasetseries/%4s/distributions",
-                        Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308")))
+                        config.getRootURL(), "common", "sample_dataset", "20230308")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseDistributionResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -189,11 +180,11 @@ public class FusionTest {
         Fusion f = stubFusion();
 
         doNothing()
-                .when(apiDownloader)
+                .when(apiManager)
                 .callAPIFileDownload(
                         String.format(
                                 "%scatalogs/%s/datasets/%s/datasetseries/%s/distributions/%s",
-                                Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308", "csv"),
+                                config.getRootURL(), "common", "sample_dataset", "20230308", "csv"),
                         String.format("%s/%s_%s_%s.%s", TMP_PATH, "common", "sample_dataset", "20230308", "csv"),
                         "common",
                         "sample_dataset");
@@ -206,11 +197,11 @@ public class FusionTest {
         Fusion f = stubFusion();
 
         doNothing()
-                .when(apiDownloader)
+                .when(apiManager)
                 .callAPIFileDownload(
                         String.format(
                                 "%scatalogs/%s/datasets/%s/datasetseries/%s/distributions/%s",
-                                Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308", "csv"),
+                                config.getRootURL(), "common", "sample_dataset", "20230308", "csv"),
                         String.format("%s/%s_%s_%s.%s", "downloads", "common", "sample_dataset", "20230308", "csv"),
                         "common",
                         "sample_dataset");
@@ -222,10 +213,10 @@ public class FusionTest {
     public void testFileDownloadAsStreamInteraction() throws Exception {
         Fusion f = stubFusion();
 
-        when(apiDownloader.callAPIFileDownload(
+        when(apiManager.callAPIFileDownload(
                         String.format(
                                 "%scatalogs/%s/datasets/%s/datasetseries/%s/distributions/%s",
-                                Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308", "csv"),
+                                config.getRootURL(), "common", "sample_dataset", "20230308", "csv"),
                         "common",
                         "sample_dataset"))
                 .thenReturn(new ByteArrayInputStream("A,B,C\nD,E,F".getBytes()));
@@ -245,11 +236,11 @@ public class FusionTest {
 
         f.upload("common", "sample_dataset", "20230308", "csv", "/tmp/file.csv", d);
 
-        verify(apiUploader)
+        verify(apiManager)
                 .callAPIFileUpload(
                         String.format(
                                 "%scatalogs/%s/datasets/%s/datasetseries/%s/distributions/%s",
-                                Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308", "csv"),
+                                config.getRootURL(), "common", "sample_dataset", "20230308", "csv"),
                         "/tmp/file.csv",
                         "common",
                         "sample_dataset",
@@ -267,11 +258,11 @@ public class FusionTest {
 
         f.upload("common", "sample_dataset", "20230308", "csv", requestBodyStream, d, d, d);
 
-        verify(apiUploader)
+        verify(apiManager)
                 .callAPIFileUpload(
                         String.format(
                                 "%scatalogs/%s/datasets/%s/datasetseries/%s/distributions/%s",
-                                Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "20230308", "csv"),
+                                config.getRootURL(), "common", "sample_dataset", "20230308", "csv"),
                         requestBodyStream,
                         "common",
                         "sample_dataset",
@@ -291,11 +282,10 @@ public class FusionTest {
 
     private Fusion stubFusion() {
         return Fusion.builder()
-                .credentials(new BearerTokenCredentials("my token"))
+                .configuration(config)
+                .bearerToken("my token")
                 .api(apiManager)
                 .responseParser(responseParser)
-                .uploader(apiUploader)
-                .downloader(apiDownloader)
                 .build();
     }
 
@@ -306,7 +296,7 @@ public class FusionTest {
         Map<String, Catalog> stubResponse = new HashMap<>();
         stubResponse.put("first", Catalog.builder().identifier("catalog1").build());
 
-        when(apiManager.callAPI(String.format("%1scatalogs", Fusion.DEFAULT_ROOT_URL)))
+        when(apiManager.callAPI(String.format("%1scatalogs", config.getRootURL())))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseCatalogResponse("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -323,7 +313,7 @@ public class FusionTest {
         catalog1.put("catalog1", Catalog.builder().identifier("catalog1").build());
         stubResponse.put("catalog1", catalog1);
 
-        when(apiManager.callAPI(String.format("%1scatalogs/common", Fusion.DEFAULT_ROOT_URL)))
+        when(apiManager.callAPI(String.format("%1scatalogs/common", config.getRootURL())))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseResourcesUntyped("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -340,8 +330,8 @@ public class FusionTest {
         dataset1.put("dataset1", Dataset.builder().identifier("dataset1").build());
         stubResponse.put("dataset1", dataset1);
 
-        when(apiManager.callAPI(String.format(
-                        "%1scatalogs/%2s/datasets/%3s", Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset")))
+        when(apiManager.callAPI(
+                        String.format("%1scatalogs/%2s/datasets/%3s", config.getRootURL(), "common", "sample_dataset")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseResourcesUntyped("{\"key\":value}")).thenReturn(stubResponse);
 
@@ -361,7 +351,7 @@ public class FusionTest {
 
         when(apiManager.callAPI(String.format(
                         "%1scatalogs/%2s/datasets/%3s/datasetseries/%4s",
-                        Fusion.DEFAULT_ROOT_URL, "common", "sample_dataset", "1")))
+                        config.getRootURL(), "common", "sample_dataset", "1")))
                 .thenReturn("{\"key\":value}");
         when(responseParser.parseResourcesUntyped("{\"key\":value}")).thenReturn(stubResponse);
