+++ src/test/java/io/github/jpmorganchase/fusion/api/FusionApiManagerTest.java
@@ -7,10 +7,11 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
+import io.github.jpmorganchase.fusion.FusionInitialisationException;
 import io.github.jpmorganchase.fusion.api.exception.APICallException;
 import io.github.jpmorganchase.fusion.http.Client;
 import io.github.jpmorganchase.fusion.http.HttpResponse;
-import io.github.jpmorganchase.fusion.oauth.provider.SessionTokenProvider;
+import io.github.jpmorganchase.fusion.oauth.provider.FusionTokenProvider;
 import java.net.HttpURLConnection;
 import java.util.HashMap;
 import java.util.Map;
@@ -29,13 +30,13 @@ public class FusionApiManagerTest {
     private Client client;
 
     @Mock
-    private SessionTokenProvider sessionTokenProvider;
+    private FusionTokenProvider fusionTokenProvider;
 
     private String apiPath;
 
     private final Map<String, String> requestHeaders = new HashMap<>();
 
-    private APICallException thrown;
+    private Throwable thrown;
 
     private String responseBody;
 
@@ -54,7 +55,7 @@ public class FusionApiManagerTest {
     }
 
     private void givenSessionBearerToken(String token) {
-        given(sessionTokenProvider.getSessionBearerToken()).willReturn(token);
+        given(fusionTokenProvider.getSessionBearerToken()).willReturn(token);
     }
 
     @Test
@@ -64,8 +65,17 @@ public class FusionApiManagerTest {
         givenSessionBearerToken("my-token");
         givenRequestHeader("Authorization", "Bearer my-token");
         givenCallToClientToGetReturnsNotFound();
-        whenFusionApiManagerIsCalledThenExceptionShouldBeThrown();
+        whenFusionApiManagerIsCalledThenExceptionShouldBeThrown(APICallException.class);
         thenExceptionMessageShouldMatchExpected("The requested resource does not exist.");
+    }
+
+    @Test
+    public void constructionWithNoCredentialsThrowsException() {
+        FusionInitialisationException thrown = assertThrows(
+                FusionInitialisationException.class,
+                () -> FusionAPIManager.builder().build(),
+                "Expected FusionInitialisationException but none thrown");
+        assertThat(thrown.getMessage(), is(equalTo("No Fusion credentials provided, cannot build Fusion instance")));
     }
 
     private void thenTheResponseBodyShouldMatchExpected() {
@@ -92,11 +102,9 @@ public class FusionApiManagerTest {
         assertThat(thrown.getMessage(), is(equalTo(message)));
     }
 
-    private void whenFusionApiManagerIsCalledThenExceptionShouldBeThrown() {
+    private void whenFusionApiManagerIsCalledThenExceptionShouldBeThrown(Class<? extends Throwable> exceptionClass) {
         thrown = assertThrows(
-                APICallException.class,
-                () -> fusionAPIManager.callAPI(apiPath),
-                "Expected APICallException but none thrown");
+                exceptionClass, () -> fusionAPIManager.callAPI(apiPath), "Expected Exception but none thrown");
     }
 
     private void givenCallToClientToGetReturnsNotFound() {
@@ -115,6 +123,9 @@ public class FusionApiManagerTest {
     }
 
     private void givenFusionApiManager() {
-        fusionAPIManager = new FusionAPIManager(client, sessionTokenProvider);
+        fusionAPIManager = FusionAPIManager.builder()
+                .httpClient(client)
+                .tokenProvider(fusionTokenProvider)
+                .build();
     }
 }
